Extract private VerifyLog method to partial class
This separates the logic of verifying logs and the monotonous public extension method overloads.
